% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_data_fetch.R
\name{cc_data_fetch}
\alias{cc_data_fetch}
\title{Fetch data}
\usage{
cc_data_fetch(key, overwrite = FALSE)
}
\arguments{
\item{key}{A S3 object, either a character string, or a \code{s3_object}
class}

\item{overwrite}{(logical) Whether to overwrite files if they already
exist on your machine. Default: \code{FALSE}}
}
\value{
A character vector of full file paths. A print method makes a tidy
return object
}
\description{
Fetch data
}
\details{
Note that data is not read into R as data can be very large.
See \code{\link{cc_data_read}}
}
\examples{
\dontrun{
key <- paste0(
 "ccafs/ccafs-climate/data/ipcc_5ar_ciat_downscaled/rcp2_6/2030s/bcc_csm1_1_m/10min/",
 "bcc_csm1_1_m_rcp2_6_2030s_prec_10min_r1i1p1_no_tile_asc.zip")
(res <- cc_data_fetch(key = key))
# indexing maintains class for easier subsetting
res[1]
res[[1]]
res[1:2]
}
}

