% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_data_read.R
\name{cc_data_read}
\alias{cc_data_read}
\title{Read CCAFS data into R}
\usage{
cc_data_read(x)
}
\arguments{
\item{x}{A \code{ccafs_files} object, the output from a call to
\code{\link{cc_data_fetch}}}
}
\value{
\code{RasterLayer} or \code{RasterBrick} class object
}
\description{
Read CCAFS data into R
}
\details{
Look in \code{rappdirs::user_cache_dir("ccafs")} for what files
are cached and to delete any.
}
\examples{
\dontrun{
key <- "ccafs/ccafs-climate/data/ipcc_5ar_ciat_downscaled/rcp2_6/
 2030s/bcc_csm1_1_m/10min/
 bcc_csm1_1_m_rcp2_6_2030s_prec_10min_r1i1p1_no_tile_asc.zip"
res <- cc_data_fetch(key = key)

# a single file
cc_data_read(res[1])

# select individual files
cc_data_read(res[1:2])

# all files
cc_data_read(res)

# character path input
## you can also pass in a path to a file(s)
cc_data_read(unclass(res[1]))

# plot data
library(raster)
plot(cc_data_read(res[1:3]))
}
}

