% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_search.R
\name{cc_search}
\alias{cc_search}
\title{Search CCAFS data}
\usage{
cc_search(file_set = NULL, scenario = NULL, model = NULL, extent = NULL,
  format = NULL, period = NULL, variable = NULL, resolution = NULL)
}
\arguments{
\item{file_set}{(integer) a file set, 2 through 12}

\item{scenario}{(integer) a scenario, 1 through 10}

\item{model}{(integer) a model, 1 through 89}

\item{extent}{(character) an extent, global (1) or region (2)}

\item{format}{(character) a format, ascii (1) or esri (2)}

\item{period}{(integer) a period, 1 through 10}

\item{variable}{(integer) a variable, 1 through 7, or 9999}

\item{resolution}{(integer) a resolutions, 1 through 7}
}
\description{
Search CCAFS data
}
\details{
See \code{\link{ccafs-search}} for details on parameters
}
\examples{
cc_search(file_set = 4, scenario = 6, model = 2, extent = "global",
  format = "ascii", period = 5, variable = 2, resolution = 3
)

cc_search(file_set = 12, extent = "global", format = "ascii",
  period = 4, variable = 1, resolution = 4
)

# a deprecated file
# cc_search(file_set = 3, scenario = 2, model = 4, extent = "global",
#   format = "ascii", period = 7, variable = 2, resolution = 5
# )
}

