% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccafs-package.R
\docType{package}
\name{ccafs-package}
\alias{ccafs}
\alias{ccafs-package}
\title{Client for CCAFS GCM Data}
\description{
Client for Climate Change, Agriculture, and Food Security
(CCAFS) General Circulation Models (GCM) data. Data is stored in
Amazon S3, from which we provide functions to fetch data.
}
\section{Vignette}{

You can access the vignette like \code{vignette(package = "ccafs")} or
\code{vignette("ccafs_vignette", package = "ccafs")}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

