% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{data_read}
\alias{data_read}
\title{Reaa data into R}
\usage{
data_read(x)
}
\arguments{
\item{x}{A \code{ccafs_files} object, the output from a call to \code{\link{data_fetch}}}
}
\value{
\code{RasterLayer} or \code{RasterBrick} class object
}
\description{
Reaa data into R
}
\examples{
\dontrun{
key <- paste0(
 "ccafs/ccafs-climate/data/ipcc_5ar_ciat_downscaled/rcp2_6/2030s/bcc_csm1_1_m/10min/",
 "bcc_csm1_1_m_rcp2_6_2030s_prec_10min_r1i1p1_no_tile_asc.zip")
res <- data_fetch(key = key)

# a single file
data_read(res[1])

# select individual files
data_read(res[1:2])

# all files
data_read(res)

# character path input
## you can also pass in a path to a file(s)
data_read(unclass(res[1]))

# plot data
plot(data_read(res[1:3]))
}
}

