% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_keys.R
\name{list_keys}
\alias{list_keys}
\title{List S3 keys}
\usage{
list_keys(prefix = NULL, delimiter = NULL, max = 1000, marker = NULL,
  ...)
}
\arguments{
\item{prefix}{Character string that limits the response to keys that begin
with the specified prefix}

\item{delimiter}{Character string used to group keys.  Read the AWS doc for more detail.}

\item{max}{Integer indicating the maximum number of keys to return (max 1000).}

\item{marker}{Character string that pecifies the key to start with when
listing objects in a bucket. Amazon S3 returns object keys in alphabetical order,
starting with key after the marker in order.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A tibble (a data.frame, basically)
}
\description{
List S3 keys
}
\examples{
\dontrun{
list_keys()
list_keys(max = 10)
list_keys(max = 1)
list_keys(prefix = "ccafs/ccafs-climate/data/ipcc_5ar_ciat_downscaled/")
}
}

