% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccafs-package.R
\docType{package}
\name{ccafs-package}
\alias{ccafs}
\alias{ccafs-package}
\title{Client for CCAFS GCM Data}
\description{
Client for CCAFS GCM Data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

